Player Input and Basic Prefab Creation
Created a player prefab for a first person player controller. This primarily takes care of movement, but contains bones for the Interact action, as well as a potential "Hold" action (think picking up something in Portal).

This uses Unity's current Input System and disables the Legacy system to avoid confusion.