Interaction Reticle
The Reticle in the middle of the player's screen is normally red, but turns green when the interaction raycast encounters something with a script with the class "IInteractable" with an "IInteractable.Interact()" function.

Known bug: in it's current implementation, something in the Pickup Inventory script prevents the reticle from turning back to red after dropping (i think) the red key I used to test the reticle.