MainManager + Progress Tracker
Created a MainManager object + script to allow for progress to be tracked across scenes. In the Start() function of the MainManager, add <string, bool> key value pairs to the ProgressTracker dictionary, set them to false. when a puzzle is completed/progress is made, set the value based on the key as true. i have done this for the first puzzle in the prototype level as an example.